refactor: (#351) Next example config
