+++ examples/next/getting-started/config.js
