+++ examples/next/getting-started/src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../../faust.config';
+import '../../config';
 import { HeadlessProvider } from '@faustjs/next';
 import 'normalize.css/normalize.css';
 import React from 'react';
